==> stock-ai/01_install.ipynb <==
pip install requests

import requests
print(requests.__version__)

pip install beautifulsoup4

import bs4
print(bs4.__version__)

pip install pymongo

import pymongo
print(pymongo.__version__)

pip install pandas

import pandas
print(pandas.__version__)

pip install tensorflow

import tensorflow
print(tensorflow.__version__)

pip install newspaper3k

import newspaper
print(newspaper.__version__)

pip install matplotlib

import matplotlib
print(matplotlib.__version__)

pip install sklearn

import sklearn
print(sklearn.__version__)
