==> stock-ai/03_stock_symbols.ipynb <==
import requests
import bs4 as bs
import pandas as pd

# url = "https://en.wikipedia.org/wiki/Dow_Jones_Industrial_Average"
url = "https://en.wikipedia.org/wiki/List_of_S%26P_500_companies"
resp = requests.get(url)
# print(resp.text)

soup = bs.BeautifulSoup(resp.text, 'lxml')
table = soup.find("table", {"class": "wikitable sortable"})

symbols = []
names = []
industries = []

for row in table.findAll('tr')[1:]:
    symbols.append(row.findAll('td')[0].text.strip())
    names.append(row.findAll('td')[1].text.strip())
    industries.append(row.findAll('td')[3].text.strip())

print(symbols)
print(names)
print(industries)

df = pd.DataFrame(list(zip(symbols, names, industries)), columns = ['symbol', 'name', 'industry'])
df
