==> stock-ai/04_stock_price_scraper.ipynb <==
import requests
from datetime import datetime
import pandas as pd

def scrape_price_data(symbol, start_date):
    end_date = int(datetime.now().timestamp())
    interval = "1d"
    url = 'https://query1.finance.yahoo.com/v7/finance/download/%s?period1=%s&period2=%s&interval=%s&events=history' %(symbol, start_date, end_date, interval)
    
    print(url)
    
    resp = requests.get(url)
    return resp.text
# scrape_price_data("AMZN", 0)   

def get_price_data(symbol, start_date):
    data = scrape_price_data(symbol, start_date)
    with open("price_%s.csv" %(symbol), 'w') as f:
        f.write(data)
    

symbols = ['FB', 'AMZN', 'AAPL', 'NFLX', 'GOOG']
start_date = 0

for symbol in symbols:
    get_price_data(symbol, start_date)
