==> stock-ai/06_numpy.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plot

#1.vector $ Matrix
a = 5
b = np.array([1,2,3])
c = np.matrix([[1,2,3], [4,5,6]])

print(a)
print(b)
print(c)

print(type(b))
print(type(c))

print(b.shape) #3row 1col

print(c.shape) #2row 3col

print(b[2])

a = np.zeros((2,2))
a

np.ones((3,2))

np.full((3,3), 7)

np.eye(3)

np.random.random((3,3)) #0~1 random numbers

# 2, Solving the equation
m = np.array([1, -1,1,1])
m = np.reshape(m, (2,2)) #reshape - 꼭 reshape해야 하나???
print(m)

m2 = np.linalg.inv(m) #linalg = liniar algibra
print(m2)

b = np.array([-1,3])
answer = np.matmul(m2, b) #matmul = matrix multiplication
print(answer)

m3 = np.array([1, -1,1,1])
m3 = np.reshape(m, (2,2)) #reshape
b = np.array([-1,3])
answer = np.matmul(m3, b) #matmul = matrix multiplication
print(answer)

#3. Array indexing
a = np.array([[1,2,3,4],[5,6,7,8],[9,10,11,12]])
print(a)

a.ndim # ndim = number of demension

c = a[1:, 2:]
print(c)

#4. Vector Operation
v = [3,4]
w = [1,2]
print(np.linalg.norm(v))
v.dot(w)
np.dot(v,w)

v = np.append(v, [0])
w = np.append(w, [0])
print(np.cross(v,w))

# 5. Matrix Operaton
M = np.array([[1,2],[3,4]])
print(M)
print(M.ndim)
print(M.shape)

# 3-dimetional matrix
D = np.array([[[1,2],[3,4]],[[5,6],[7,8]]])
print(D)
print(D.ndim)
print(D.shape)

m = np.array([[1,2],[3,4]])
v = np.array([[3,1],[1,3]])

print(np.matmul(m,v))

q = np.array([1,2,3,4,5,6,7,8,9])
q = q.reshape(3,3)
print(q)
print(q.T)

print(np.linalg.det(q)) #det

print(np.linalg.inv(q)) #inv
