==> stock-ai/07_stock_predictor.ipynb <==
from datetime import datetime
from datetime import timedelta

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

import tensorflow as tf
from tensorflow import keras

#1. Data
df = pd.read_csv("price_GOOG.CSV")
df

# plt.plot(df["Date"].values, df["Adj Close"].values)  #pandas - loading data
# plt.show()

training_set = df.iloc[:, 5:6].values
training_set


print(training_set.ndim)
print(training_set.shape)

from sklearn.preprocessing import MinMaxScaler
sc = MinMaxScaler(feature_range=(0,1))
training_set_scaled = sc.fit_transform(training_set)
print(training_set_scaled)

X_train = [] #past 60'd price
Y_train = [] #today's price

for i in range(60, training_set_scaled.shape[0]):
    X_train.append(training_set_scaled[i-60:i, 0])
    Y_train.append(training_set_scaled[i,0])
    
X_train = np.array(X_train)
Y_train = np.array(Y_train)

print(X_train)
print(X_train.shape)
print(X_train.ndim)
print(Y_train)
print(Y_train.shape)

X_train = np.reshape(X_train, (X_train.shape[0], X_train.shape[1] ,1))
print(X_train)
print(X_train.ndim)
print(X_train.shape)

# 2.Model & Traning

model = keras.Sequential([
    keras.layers.LSTM(units=50, return_sequences=True, input_shape=(X_train.shape[1],1)),
    keras.layers.Dropout(0.2),
    keras.layers.LSTM(units=50, return_sequences=True),
    keras.layers.Dropout(0.2),
    keras.layers.LSTM(units=50, return_sequences=True),
    keras.layers.Dropout(0.2),
    keras.layers.LSTM(units=50),
    keras.layers.Dropout(0.2),
    keras.layers.Dense(units=1)
])

model.compile(
    optimizer = 'adam',
    loss = 'mse'  
)

model.summary()

# model.fit(X_train, Y_train, epochs = 100)
model.fit(X_train, Y_train, epochs = 10)

#3.preict Data
print(df.head())
real_stock_data = df.iloc[-60:, 5:6].value
print(real_stock_data)

X_test = []
for i in range(0, 60):
    val = df.iloc[-60-60+i:-60+i, 5:6].values
    val = sc.transform(val)
    X_test.append(val)
    
X_test = np.array(X_test)
X_test = X_test.reshape(X_test.shape[0], X_test.shape[1],1)

print("X_test.shape")
print(X_test.shape)
print(X_test[:1])

predicted = model.predict(X_test)
predicted = sc.inverse_transform(predicted)

plt.plot(real_stock_data, color='red', label = 'real_stock_data')
plt.plot(predicted, color="blue", label = "Predicted Stock price")
plt.xlabel("date")
