==> stock-ai/09_stock_article_dataframe_saver.ipynb <==
import sys
import pandas as pd
import traceback #excepiton
import glob #Reading the file
import json

def find_json_files(symbol):
    try:
        return glob.glob("articles/*.json")
    except Exception as ex:
        print(ex)
        tractback.print_exc(file=sys.stdout)

# find_json_files("AMZN")

def convert(file):
    with open(file) as f:
        ad = json.load(f)
    #print(ad)

    row = [{
        'date': ad['publish_date'],
        'symbol': ad['symbol'],
        'url': ad['link'],
        'raw': ad['text']
    }]
    
    return row

# convert("articles/NFLX-2020-05-08T12:15:08+00:00.json")

def make_articles_dataframe(symbol):
    df = pd.DataFrame(columns=['url', 'symbol', 'date', 'raw' ])
    
    files = find_json_files(symbol)
    for file in files:
        row = convert(file)
        df = df.append(row, ignore_index = True)
        
    return df

def main():
    df = make_articles_dataframe("AMZN")
    df.to_csv("AMZN_STEP1.csv", index=False)
    return df

main()

#NLP
