==> stock-ai/10_stock_article_prep.ipynb <==
import nltk
nltk.download('wordnet')
nltk.download('stopwords')

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer

import string

#pip install nltk

print(nltk.__version__)

#pip install sklearn

import sklearn
print(sklearn.__version__)

print(string.punctuation)

print(nltk.corpus.stopwords.words('english'))

a = 'HELLO World'
print(a.lower())

# steming => stem() : find root word
ps = nltk.PorterStemmer()

print(ps.stem('meaning'))
print(ps.stem('electricity'))
print(ps.stem('stemming'))
print(ps.stem('stemmed'))
print(ps.stem('grows'))
print(ps.stem('growing'))
print(ps.stem('grow'))
print(ps.stem('type'))
print(ps.stem('typed'))
print(ps.stem('typing'))

# WordNetLemmatizer
wn = nltk.WordNetLemmatizer()

print(wn.lemmatize('geese'))
print(wn.lemmatize('goose'))
print(wn.lemmatize('foot'))
print(wn.lemmatize('feet'))
