==> stock-ai/11_stock_article_predictor.ipynb <==
import pandas as pd
#Import scikit-learn metrics module for accuracy calculation

from sklearn.model_selection import train_test_split
from sklearn import metrics
from sklearn.ensemble import RandomForestClassifier

def prepare_testing_set(df, percent_testing):
    X = df.drop("@@label", axis=1)
    Y = df["@@label"]
    
    return train_test_split(X, Y, test_size = percent_testing)
    

def main():
    df = pd.read_csv("AMZN_STEP2.csv")
    df.drop(columns = ["@@date", "@@url"], inplace=True)
    
    X_train, X_test, y_train, y_test = prepare_testing_set(df, 0.2)
    
    model = RandomForestClassifier(n_estimators=4000, max_depth=70).fit(X_train, y_train)
    y_pred = model.predict(X_test)
    
    print(X_test[:3])
    print(list(y_test))
    print(list(y_pred))
    
    print("Accuracy: ", metrics.accuracy_score(y_test, y_pred))
    
main()
